frontend/requirements.txt
fastapi
uvicorn
google-generativeai
langchain
langchain-google-genai
python-multipart
python-dotenv
PyPDF2
mangum